Add beyas classifier and get result for tomorrow price using classify method

+++ app/src/main/kotlin/com/farahani/elmira/bitcoinpredictor/utils/mapper.kt
@@ -8,6 +8,6 @@ fun Dto.BitcoinHistory.map()= BitcoinHistoryModel(
         .replace("{", "")
         .replace("}", "")
         .split(",").map {
-            it.split("=")[1]
+            it.split("=")[1].toDouble()
         }
 )
